refactor(vscode): lift frame carousel out of behavior.ts
Slice of #767. Extracts the per-card frame carousel for animated /
multi-capture previews — `buildFrameControls`, `stepFrame`, `showFrame`,
`updateFrameIndicator` — into a typed `FrameCarouselController` class.

The shared `cardCaptures` Map (populated from `setPreviews`, mutated
by `updateImage` / `setImageError`, drained on card removal) stays in
`behavior.ts` and is handed to the controller via constructor. Same
DOM shape (`.frame-controls`, `.frame-indicator`, `.frame-prev`,
`.frame-next`), same arrow-key handling, same per-capture image swap
and error-panel attach.

Drive-by: fixes a latent bug in the no-imageData-but-error branch of
`showFrame` that was calling `buildErrorPanel(message, renderError,
className)` instead of the typed signature
`buildErrorPanel(vscode, message, renderError, className)` introduced
in #770. The other call site in the same function had it right; this
one was missed and `@ts-nocheck` on `behavior.ts` hid it. Caught
when copying the body into the typed module.

`behavior.ts` shrinks from 2217 to 2085 lines (-132 LOC).

Refs #767.
